refactor: combine all build and release workflows
Merge the build-guest, build-host and release workflows into a single one.
The matrix now runs by distro. Each job builds the guest image first, then
the host image embeds it since they now share a build directory. This
removes the need to build the guest image twice and significantly
reduces overall build time.

A devel tag is generated and new images are uploaded only when code is
merged into main. This removes the need to upload and download
artifacts. The build still triggers in PRs which helps us catch errors
in changes to the mkosi process.

When a new tag is created, it will automatically create a release for
the tag and upload the images for that tag.

Signed-off-by: DGonzalezVillal <Diego.GonzalezVillalobos@amd.com>